refactor: Renames cofundable.routes to cofundable.routers
