+++ src/cofundable/api.py
@@ -3,7 +3,7 @@
 from fastapi import FastAPI
 from fastapi_pagination import add_pagination
 
-from cofundable.routes.causes import cause_router
+from cofundable.routers.causes import cause_router
 
 app = FastAPI()
 app.include_router(cause_router)
